requirements.txt
discord.py==2.4.0
flask==3.0.3
python-dotenv==1.0.1
